additional pruning
