+++ scripts/linters/find_banned_deps.rb
@@ -1,39 +1,53 @@
 # frozen_string_literal: true
 
-# This script checks all `Cargo.toml` files for banned dependencies, fails
-# the GitHub Continuous Integration dependencies check if any banned dependencies
-# are found, and reports any violations in the Continuous Integration output. If
-# additional exceptions need to be made, please add these to the `whitelist`.
+# This script checks all `Cargo.toml` and `*.rs` files for banned dependencies,
+# fails the GitHub Continuous Integration dependencies check if any banned
+# dependencies are found, and reports any violations in the Continuous
+# Integration output. If additional exceptions need to be made, please add
+# these to the appropriate list (i.e., `toml_list` for `Cargo.toml` files and
+# `rs_list` for Rust source code files).
 
 require 'toml-rb'
 
-whitelist = {
+toml_list = {
   # Check all `Cargo.toml` files for `time` crate
   'time' => []
+}
+rs_list = {
   # Create exceptions for `forest_shim` and `forest_interpreter` (only places
   # `fvm` & `fvm3` are allowed).
-  #
-  # Fixup in #2991
-  #
-  # 'fvm' => ['utils/forest_shim/Cargo.toml', 'vm/interpreter/Cargo.toml'],
-  # 'fvm3' => ['utils/forest_shim/Cargo.toml', 'vm/interpreter/Cargo.toml']
+  'fvm' => ['src/interpreter', 'src/shim'],
+  'fvm3' => ['src/interpreter', 'src/shim'],
+  'fvm_shared' => ['src/interpreter', 'src/shim', 'src/json/tests/'],
+  'fvm_shared3' => ['src/interpreter', 'src/shim']
 }
 
 violations = Hash.new { |h, k| h[k] = [] }
 
-# Iterate over all TOML files in the repository
+# Iterate over TOML file(s) in the repository
 Dir.glob('**/*.toml').each do |file|
   toml = TomlRB.load_file(file)
 
   toml['dependencies']&.each do |dep|
-    whitelist.each_pair do |crate, exceptions|
+    toml_list.each_pair do |crate, exceptions|
       violations[dep[0]] << file if dep[0] == crate && !(exceptions.include? file)
     end
   end
 
   toml['dev-dependencies']&.each do |dep|
-    whitelist.each_pair do |crate, exceptions|
+    toml_list.each_pair do |crate, exceptions|
       violations[dep[0]] << file if dep[0] == crate && !(exceptions.include? file)
+    end
+  end
+end
+
+# Iterate over all Rust source code files in the repository
+Dir.glob('**/*.rs').each do |filename|
+  File.foreach(filename) do |file|
+    rs_list.each_pair do |crate, exceptions|
+      pattern = "#{crate}::"
+      match = file.match pattern
+      violations[crate] << filename if !(exceptions.any? { |exception| filename.include? exception }) && !match.nil?
     end
   end
 end
