+++ src/json/sector.rs
@@ -2,10 +2,10 @@
 // SPDX-License-Identifier: Apache-2.0, MIT
 
 pub mod json {
+    use crate::shim::sector::SectorNumber;
     use crate::shim::sector::{PoStProof, RegisteredPoStProof, RegisteredSealProof, SectorInfo};
     use base64::{prelude::BASE64_STANDARD, Engine};
     use cid::Cid;
-    use crate::shim::sector::SectorNumber;
     use serde::{de, Deserialize, Deserializer, Serialize, Serializer};
 
     /// Wrapper for serializing a `PoStProof` to JSON.
