+++ src/message/signed_message.rs
@@ -1,6 +1,7 @@
 // Copyright 2019-2023 ChainSafe Systems
 // SPDX-License-Identifier: Apache-2.0, MIT
 
+use crate::shim::message::MethodNum;
 use crate::shim::{
     address::Address,
     crypto::{Signature, SignatureType},
@@ -9,7 +10,6 @@ use crate::shim::{
 };
 use fvm_ipld_encoding::{to_vec, Cbor, Error as CborError};
 use fvm_ipld_encoding3::RawBytes;
-use fvm_shared::MethodNum;
 use serde_tuple::{self, Deserialize_tuple, Serialize_tuple};
 
 use super::Message as MessageTrait;
