+++ src/shim/clock.rs
@@ -1,8 +1,8 @@
 // Copyright 2019-2023 ChainSafe Systems
 // SPDX-License-Identifier: Apache-2.0, MIT
 
-pub use fvm_shared::BLOCKS_PER_EPOCH;
 pub use fvm_shared::ALLOWABLE_CLOCK_DRIFT;
+pub use fvm_shared::BLOCKS_PER_EPOCH;
 
 pub const SECONDS_IN_DAY: i64 = 86400;
 pub const EPOCHS_IN_DAY: i64 = SECONDS_IN_DAY / EPOCH_DURATION_SECONDS;
